third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/text-combine-upright-parsing-valid-001-expected.txt

This is a testharness.js-based test.
FAIL Computed value for `text-combine-upright: none` is `none` assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: all` is `all` assert_equals: expected (string) "all" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: initial` is `none` assert_equals: expected (string) "none" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: inherit` is `all` which is inherited from the parent element assert_equals: expected (string) "all" but got (undefined) undefined
FAIL Computed value for `text-combine-upright: unset` is `all` which is inherited from the parent element assert_equals: expected (string) "all" but got (undefined) undefined
Harness: the test ran to completion.
